New CLI Features & Refactor - Added new types (CommitConfig, GitChanges, OllamaResponse, ModelInfo) - Implemented version info - Removed old index.ts file - Modified src/cli/index.ts to include the new features ( new functions/vars) - Now providing more context & structure for your commit messages
